refactor: give Connect component clearer name and reorganise helpers
