+++ pages/api/stats.tsx
@@ -1,7 +1,7 @@
 import { withSentry } from '@sentry/nextjs'
 import { NextApiRequest, NextApiResponse } from 'next'
 
-import { cacheObject } from '../../helpers/cacheObject'
+import { cacheObject } from '../../helpers/api/cacheObject'
 import { getOasisStats } from '../../server/services/getOasisStats'
 
 const getStats = cacheObject(getOasisStats, 12 * 60 * 60)
